refactor(reid): keep shared tracker utilities reusable
