+++ src/trackers/core/reid/__init__.py
@@ -10,12 +10,14 @@ import importlib
 
 __all__ = [
     "FASTREID_MOT17_SBS50",
+    "FeatureBank",
     "ReIDEvaluator",
     "ReIDMetrics",
     "ReIDModel",
     "ReIDPreprocessing",
     "ReIDResult",
     "ReIDSplit",
+    "appearance_similarity",
     "compute_reid_metrics",
     "load_market1501",
     "load_msmt17",
@@ -38,9 +40,11 @@ _REID_OPTIONAL_ROOTS = frozenset(
 )
 
 # NumPy-only symbols — safe to import without torch/timm/HF.
+from trackers.core.reid.appearance import appearance_similarity
 from trackers.core.reid.eval.datasets import ReIDSplit, load_market1501, load_msmt17
 from trackers.core.reid.eval.evaluator import ReIDEvaluator, ReIDResult
 from trackers.core.reid.eval.metrics import ReIDMetrics, compute_reid_metrics
+from trackers.core.reid.feature_bank import FeatureBank
 
 _LAZY_EXPORTS: dict[str, tuple[str, str]] = {
     "ReIDModel": ("trackers.core.reid.model", "ReIDModel"),
